# PERF: DataFrame initialisation with dict with single key/value pair