Faster initialisation of DataFrames, when the data is a single-item dict:

```python
>>> x = np.arange(1_000_000)
>>> %timeit pd.DataFrame(dict(a=x))
2.16 ms ± 66.7 µs per loop  # master
282 µs ± 4.63 µs per loop  # this PR
>>> x = np.array([str(u) for u in range(1_000_000)], dtype=object)
>>> %timeit pd.DataFrame(dict(a=x))
29.5 ms ± 131 µs per loop  # master
12.6 ms ± 55.8 µs per loop  # this PR
```

I don't think it's possible to optimize the case with more than 1 key/value pair, as concatenation will always take some time to perform.